Swapped custom error level on failure in checkedKeyRetrieval to setter
- Instead of setting a custom error level on failure through
checkedKeyRetrieval, Qx::GenericError's setErrorLevel() now returns a
reference to instance the function is invoked through so that the error
instance can easily be returned elsewhere with a modified error level.
This follows the original intention of having mErrorLevel be the only
mutable member of Qx::GenericError. It was always intended for the user
to change the error level after a more general function in the Qx
library constructs a Qx::GenericError instance since the library can't
know the severity of an error without context of the main application
most of the time.