Remove endianness from BitArray
Bit manipulation in C++ is always performed from a Big Endian
persepctive (i.e. when shifting a number larger than 1 byte the bits are
ordered as if the number was Big Endian, even on a Little Endian system)
and so the bits in this class should always be read, written, and stored
in Big Endian order (when concering enough bits to make up more than 1
byte).

Add function BitArray::toByteArray() that takes endianess to make up for
the above change, since endianess does matter and changing it may be
desriable when in the context of full bytes instead of bits.