Disable Index related overloads to require more explicit syntax
While the original operator overloads for Index were fairly functional,
there were still some cases where the myriad of exceptions in
comparison/arithmetic rules set by these overloads could result is non-
obvious user code mistakes, or undesirable implicit conversions.

Because of this they were all commented out (so they can always be
restored later) and instead an overload of the indirection operator was
introduced that simply returns a constant reference to the Index's
mValue member. This allows for the value to be used in cases where
arithmetic or comparison with T or other primitives is desired with a
low syntactic footprint, but still requires the user to explicitly
access the value, essentially acknowledging that they (should) have
checked if the Index is null or Last before doing so.

Also removed the overload for lengthOfRange that took Index<T> since its
output could be nonsensical with a null/last Index.