Remove qx-globals "group" since groups must be components (modules)
Unlike QDoc, Doxygen doesn't have a seperate grouping and modules
mechanic, with them instead being one in the same.

Since the modules of Doxygen need to be used to describe Qx Components,
the "globals" group cannot be documented using that mechanism, and
instead must be left out or documented some other way.