Add boilerplate documentation for primary component headers
Because Doxygen currently has no way to merge file paths in the file
list that have the same hierarchy, this adds an extra entry to the file
list that repeats the main structure but under the build folder.

The resultant view is incredibly clunky and therefore this change isn't
considered worth the detriment and therefore is being shelved for now.