Remove connect asserts since they're generally disabled in release mode
This was preventing the connections from working in release mode as
assert expressions are not evaluated in release mode at all when NDEBUG
is defined, which was the case in this files current include setup.
Qt5s "new" connection syntax compile time checks, and the general
runtime warnings should be sufficient anyway.

Oops.