Revert to using static megatoools (don't pull unstable dependencies)
Installing from debian experimental/unstable causes newer (unstable)
versions of dependencies to be pulled (i.e. libc) which makes the
binaries from builds on that system following the install of megatools
rely on these newer versions, invalidating the purpose of using an
older, more compatible build of Ubuntu.

Annoyingly the megatools site doesn't seem to keep old builds around so
the package name will likely have to be updated frequently.

Another option is trying to build from unstable source and ensure that
the unstable repo isn't prefered by apt, but this in itself is
complicated.

Eventually when moving to Ubuntu 22.04 hopefully megatools 1.11 will be
on the stable branch. 1.10 currently is but it just outright doesn't
work.