Fix deadlock when all remaining downloads fail due to IO error
When all active tasks had finished and the only remaining ones all
failed to start due to an IO error during the call to startDownload, the
manager would enter a deadlock as this condition was not properly
checked for, so finish() was never called.

This also changes the approach of initiating downloads/enumerations
and then continuing them when needed to a cleaner one with more
convergent execution flow, less redundancy, and clear checks that handle
either starting the next download/enumeration, begin the next step, or
end processing as required.

This also constitutes a minor paradigm shift in the way that subsequent
tasks after the initial batch are handled. Previously, tasks would be
started up to the limit specified by mMaxSimultaneous and then each
completed task would trigger one more to start if the queue wasnt empty.
This would ensure that most of the time that as many tasks as allowed
were always running (minus one, plus one), but technically if multiple
downloads finished while the finished handler was still starting just
one download, the active task count could end up averaging slightly
below the allowed maximum. This would continue to be the case if the
rate at which active tasks finished eclipsed the rate that new ones were
started, however unlikely that is.

The new implementation of task initiation explicitly ensures that the
active task bin is filled to capacity each time a completed task
triggers new ones to start. In practice this should not result in a
perceptible difference, but is worth noting.