Change IoOpReport::wasSuccessful() to ::isFailure() for clearer paradigm
Now the function returns a boolean from the perspective of looking for
an error by default, instead of requiring a negation to check for an
error.

This way a null IoOpReport can reasonable be seen as neither a success
nor a failure, and instead used in contexts where nothing has occured
(i.e. Qx streamers).

To clarify, calling `.isFailure()` on a null IoOpReport returning
`false` is reasonable given the above, where as before checking !
wasSuccessful more so gave the implication that the op failed outright.
It's reasonable to consider both successful and null reports as "non-
failures" under the new perspective.