Queue the start of DownloadManager tasks to the event loop
In a somewhat blind attempt to fix the issue of DownloadManager
seemingly randomly deadlocking after completing size enumeration, use a
singleshot timer to queue the start of all tasks (enumeration or
download) until the next event loop iteration, instead of starting them
immediately. Most significantly, this will ensure that no true downloads
are started after enumeration is finished until control has returned to
the event loop and the last enumeration reply has been deleted.

Aside from causing a small delay (which in the past has seemed to help
workaround the issue, e.g. when stopping at a break point before the
download starts), the potential theory with this is that
QNetworkAccessManager might behave finicky when opening a request to the
same URL as an already processed reply. This would present issues
downloading the same file multiple times within the same queue, but this
is an uncommon case at the very least.

If the issue still occurs, adding another delay by using a singleshot
timer for the call to startTrueDownloads() within
pushEnumerationsUntilFinished() could be tried.