Constrain String::join to fix overload resolution for T = QString
This constrain ensures that the function parameter for the 4 argument
versions is actually a function that takes type T as an input and
returns a QString. This is better in general, but specifically fixes
overload resolution not choosing the correct overload for when T is
already a QString (i.e. in the convenience overloads of String::join).