Fix some contents of component include dirs not being installed
This change demonstrates an important distinction within the use of the
CMake install(DIRECTORY ...) command.

Originally the public headers of each component were install like so:

`````````````````````````````````````````````````````````````````````````
install(DIRECTORY include/${PROJECT_NAME_LC}
    COMPONENT ${COMPONENT_TARGET_NAME}
    DESTINATION "include/${COMPONENT_NAME_LC}"
    ...
)
`````````````````````````````````````````````````````````````````````````

This tells CMake to install the folder (for example, lets say the 'core'
component) "include/qx" to "include/core" such that in the end there
will be "include/core/qx/...". This made it seem like CMake can only
copy the whole directory and was not able to copy just the directory's
contents; however, it turns out this can be controlled by adding a
trailing slash to the first argument, which will make CMake only copy
the contents. This is reflected in the new method:

`````````````````````````````````````````````````````````````````````````
install(DIRECTORY include/
    COMPONENT ${COMPONENT_TARGET_NAME}
    DESTINATION "include/${COMPONENT_NAME_LC}/"
    ...
)
`````````````````````````````````````````````````````````````````````````

This ensures that the entire contents of the components include
directory are copied to the install directory, instead of just the "qx"
folder within. For most components this didn't matter, but qx-windows
has an extra include outside the "qx" sub-folder (qx_windows.h) that was
not being copied because of the old approach.