Completely overhaul library error system
There are too many changes to list verbatim, but primarily:

Introduces the Qx::Error interface. The interface acts as a pseudo-
polymorphic container/"base" for specific error types that are derived
from the actual base implementation AbstractError (which itself derives
from the literal interface (virtual) class IError.

Error is able to be constructed from any class derived from
AbstractError such that instances of said classes can be used in a
polymorphic-like fashion but with value semantics. Any variable/argument
of type Qx::Error can be satisfied by an AbstractError derived type.

The interface establishes a set of common properties for all error
types:
- Type code, which is specific to each error type (read static) and
guaranteed to be unique
- Type name, usually matches the literal class name, same restrictions
as above
- Value (i.e. error code)
- Severity
- Caption
- Primary info
- Secondary info
- Details

The Error::code()/hexCode() methods produces a 64-bit value that can be
used to completely represent a specific occurrence of an error
(potentially minus some minor details) as it is composed of the error's
type, value, and severity.

Additionally, any arbitrary type not derived from AbstractError can be
adapted to work with the interface via an Error Adaptation. Essentially,
this is a specifically composed derivation of AbstractError paired with
the type to be adapted. The adapter allows that type to be implicitly
converted to an Qx::Error object once registered.

GenericError has been reworked to be derived from AbstractError and act
as a minimal way to work with the interface, though more context
specific error types should generally be preferred. Following this,
several new error types that comply with the interface have been created
within the library where GenericError was previously used.

These changes also allowed for the reduction of code duplication across
qx-system and the windows component as error handling in both cases now
utilize SystemError, which contains the implementation to convert from
underlying system specific errors.