Rewrite Index<T> comparisons using C++20 guidance of only == and <=>
This change was motivated by the introduction of P2468R2 which caused
the behavior of the languages comparison rules to fallback to C++17
behavior if a class had both operator()== and operator()!= manually
implemented. That standard isn't as flexible with implicit conversions
for operators and so compilation was failing for qx-common-io since it
relies on implicitly converting an int to an Index32 for purposes of
comparison.

Simply removing the manual inequality operator restores the original C+
+20 behavior, but while at it the other operators were changed to be
handled by the three-way comparison operator as well.