Add struct based JSON parsing
Add ability to parse arbitrary user structs via the use of declarative
macros placed within said structs that simply re-list the data members
of the struct.

By using QX_JSON_STRUCT(...) where ... is a list of QX_JSON_MEMBER macro
invocations that use the name of the structs data members, one can make
that struct directly parseable as if it as a JSON object.

The name of members must match the name of their corresponding key in
actual JSON data.

There is much to improve in terms of simplifying the macro requirements
(i.e. dropping requirement of inner macro) and adding flexibility (like
handling derived structs and allowing for members to be associated with
one), but for now it's very usable
as is.

This commit also removes the Json class, only moving a few of its
functions to the Qx namespace, as most of its functionality is made
obsolete by these new facilities.
 feature/improved_json (#113)