Add Qx Bindable Properties, an alternative to Qt Bindable Properties
- Very similar API
- Drops binding evaluations to the absolute minimum
- Allows on-the-fly pickups of new dependencies (i.e. bindings
  don't need to worry about code-paths as much)
- QOL changes to API and objects
  - No templated callback handlers
  - Lifetime callback registration option

Still missing some of the polish of the originals, plus no support for
QML (yet at least) or MOC (likely never), so there are downside, but
largely we're at feature parity, plus have improvements.

The core "algorithm"/paradigm that always keeps binding evaluation at a
minimum is the most important part and at the very least acts as a model
of how the Qt system should potentially handle things, if ever there was
a pipe dream to have that portion of this upstreamed.